Extract abstract FeedMessage which have all common fields
